hf-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
psycopg2-binary==2.9.11
python-multipart==0.0.6
passlib[bcrypt,argon2]==1.7.4
bcrypt==4.0.1
argon2-cffi==21.3.0
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
